fix: carry peer provenance into a dispatched task's opening brief
`send` prefixes a cross-task prompt with the sender and a reply command;
`add --prompt` did not, so a task started by another agent could not answer it.
It knew only what the brief said, and the dispatcher link lived on the task row
where nothing pointed at it — a survey finished and sat waiting rather than
reporting (issue #92).

Moves withPeerProvenance to dispatcher.ts (both callers already import it) and
applies it in add. Unchanged for a create from a plain shell.
